requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dateutil==2.8.2
torch==2.1.0
transformers==4.35.2
nltk==3.8.1
pandas==2.1.3
websocket-client==1.6.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
streamlit==1.28.2
lxml==4.9.3
pillow==10.1.0
